chore: some minor clean ups
