+++ package.json
@@ -48,7 +48,7 @@
     "js-yaml": "4.1.0",
     "mime": "3.0.0",
     "minimatch": "5.1.0",
-    "ramda": "^0.28.0",
+    "ramda": "0.28.0",
     "rcedit": "3.0.1",
     "semver": "7.3.7",
     "zod": "3.17.10"
@@ -61,7 +61,7 @@
     "@types/mime": "3.0.0",
     "@types/minimatch": "3.0.5",
     "@types/node": "18.6.3",
-    "@types/ramda": "^0.28.15",
+    "@types/ramda": "0.28.15",
     "@types/semver": "7.3.10",
     "@types/uuid": "8.3.4",
     "esbuild": "0.14.53",
