+++ release/src/changelogFormatters/prChangelog.ts
@@ -1,21 +1,8 @@
 import { ChangelogItem } from '../changelog'
 
-export function formatLinkType(type: string) {
-  switch (type) {
-    case 'commit':
-      return 'Commit'
-    case 'pr':
-      return 'Pull request'
-    case 'issues':
-      return 'Issue'
-    case 'milestone':
-      return 'Milestone'
-    default:
-      return 'Unknown'
-  }
-}
+import { formatLinkType } from './utils'
 
-export function formatPrChangelog(version: string, changelog: ChangelogItem[]) {
+export function formatPrChangelog(changelog: ChangelogItem[]) {
   return [
     ...buildChangelogSection(
       'New features',
