+++ release/src/changelogFormatters/publicChangelog.ts
@@ -1,19 +1,6 @@
 import { ChangelogItem } from '../changelog'
 
-export function formatLinkType(type: string) {
-  switch (type) {
-    case 'commit':
-      return 'Commit'
-    case 'pr':
-      return 'Pull request'
-    case 'issues':
-      return 'Issue'
-    case 'milestone':
-      return 'Milestone'
-    default:
-      return 'Unknown'
-  }
-}
+import { formatLinkType } from './utils'
 
 export function formatPublicChangelog(
   version: string,
